RavenDB-21192 Moved in-memory item errors list to EtlProcessStatistics
