+++ src/Raven.Server/Documents/ETL/EtlProcess.cs
@@ -1204,7 +1204,7 @@ namespace Raven.Server.Documents.ETL
                             var results = ravenEtl.Transform(new[] { ravenEtlItem }, context, new EtlStatsScope(new EtlRunStats()),
                                 new EtlProcessState { SkippedTimeSeriesDocs = new HashSet<string> { testScript.DocumentId } });
 
-                            var etlItemErrors = database.EtlErrorsStorage.ReadInMemoryItemErrorsOfProcess($"{testScript.Configuration.Name}/{testScript.Configuration.Transforms.First().Name}");
+                            var etlItemErrors = ravenEtl.Statistics.ReadInMemoryItemErrorsOfProcess($"{testScript.Configuration.Name}/{testScript.Configuration.Transforms.First().Name}");
                             
                             return new RavenEtlTestScriptResult
                             {
@@ -1270,7 +1270,7 @@ namespace Raven.Server.Documents.ETL
                                 }
                             }
                             
-                            var etlItemErrors = database.EtlErrorsStorage.ReadInMemoryItemErrorsOfProcess($"{testScript.Configuration.Name}/{testScript.Configuration.Transforms.First().Name}");
+                            var etlItemErrors = olapElt.Statistics.ReadInMemoryItemErrorsOfProcess($"{testScript.Configuration.Name}/{testScript.Configuration.Transforms.First().Name}");
 
                             return new OlapEtlTestScriptResult
                             {
